+++ examples/Console/EthanI/AI24Bit/AI_on_demand_once.cs
@@ -1,0 +1,75 @@
+/// AI_on_demand_once.cs with synchronous mode.
+///
+/// This example demonstrates the process of obtaining AI data in on demand mode.
+/// Additionally, it retrieve AI data from EthanI.
+/// For other examples please check:
+/// https://github.com/WPC-Systems-Ltd/WPC_CSharp_driver_release/tree/main/examples
+/// See README.md file to get detailed usage of this example.
+///
+/// Copyright (c) 2023 WPC Systems Ltd.
+/// All rights reserved.
+
+using WPC.Product;
+
+class EthanI_AI_on_demand_once
+{
+    static public void Main()
+    {
+        // Get C# driver version
+        Console.WriteLine($"{Const.PKG_FULL_NAME} - Version {Const.VERSION}");
+
+        // Create device handle
+        EthanI dev = new EthanI();
+
+        // Connect to device
+        try
+        {
+            dev.connect("192.168.1.110"); // Depend on your device
+        }
+        catch (Exception ex)
+        {
+            Console.WriteLine(ex);
+            // Release device handle
+            dev.close();
+            return;
+        }
+
+        try
+        {
+            // Parameters setting
+            int err;
+            int port = 0;
+            int timeout = 3000; // ms
+
+            // Get firmware model & version
+            string[] driver_info = dev.Sys_getDriverInfo(timeout:timeout);
+            Console.WriteLine($"Model name: {driver_info[0]}");
+            Console.WriteLine($"Firmware version: {driver_info.Last()}");
+
+            // Open AI
+            err = dev.AI_open(port, timeout:timeout);
+            Console.WriteLine($"AI_open in port {port}: {err}");
+
+            // Read data acquisition acquisition
+            List<float> sample = dev.AI_readOnDemand(port, timeout:timeout);
+
+            // Print data
+            Console.WriteLine(string.Format("[{0}]", string.Join(", ", sample)));
+
+            // Close AI
+            err = dev.AI_close(port, timeout:timeout);
+            Console.WriteLine($"AI_close in port {port}: {err}");
+
+        }
+        catch (Exception ex)
+        {
+            Console.WriteLine(ex);
+        }
+
+        // Disconnect device
+        dev.disconnect();
+
+        // Release device handle
+        dev.close();
+    }
+}